connlib: simplify control_protocol handling
