+++ rust/connlib/libs/tunnel/src/device_channel.rs
@@ -1,0 +1,9 @@
+#[cfg(target_family = "unix")]
+#[path = "device_channel/device_channel_unix.rs"]
+mod device_channel;
+
+#[cfg(target_family = "windows")]
+#[path = "device_channel/device_channel_win.rs"]
+mod device_channel;
+
+pub(crate) use device_channel::*;
