+++ rust/connlib/libs/tunnel/src/lib.rs
@@ -46,6 +46,7 @@ pub use webrtc::peer_connection::sdp::session_description::RTCSessionDescription
 use index::IndexLfsr;
 
 mod control_protocol;
+mod device_channel;
 mod dns;
 mod iface_handler;
 mod index;
@@ -54,27 +55,6 @@ mod peer;
 mod peer_handler;
 mod resource_sender;
 mod resource_table;
-
-#[cfg(any(target_os = "macos", target_os = "ios"))]
-#[path = "tun_darwin.rs"]
-mod tun;
-
-#[cfg(target_os = "linux")]
-#[path = "tun_linux.rs"]
-mod tun;
-
-// TODO: Android and linux are nearly identical; use a common tunnel module?
-#[cfg(target_os = "android")]
-#[path = "tun_android.rs"]
-mod tun;
-
-#[cfg(target_family = "unix")]
-#[path = "device_channel_unix.rs"]
-mod device_channel;
-
-#[cfg(target_family = "windows")]
-#[path = "device_channel_win.rs"]
-mod device_channel;
 
 const MAX_UDP_SIZE: usize = (1 << 16) - 1;
 const RESET_PACKET_COUNT_INTERVAL: Duration = Duration::from_secs(1);
@@ -179,7 +159,7 @@ pub struct Tunnel<C: ControlSignal, CB: Callbacks> {
     gateway_awaiting_connection: Mutex<HashMap<GatewayId, Vec<IpNetwork>>>,
     resources_gateways: Mutex<HashMap<ResourceId, GatewayId>>,
     webrtc_api: API,
-    resources: RwLock<ResourceTable<ResourceDescription>>,
+    resources: Arc<RwLock<ResourceTable<ResourceDescription>>>,
     control_signaler: C,
     gateway_public_keys: Mutex<HashMap<GatewayId, PublicKey>>,
     callbacks: CallbackErrorFacade<CB>,
@@ -263,7 +243,7 @@ where
         let peers_by_ip = RwLock::new(IpNetworkTable::new());
         let next_index = Default::default();
         let peer_connections = Default::default();
-        let resources = Default::default();
+        let resources: Arc<RwLock<ResourceTable<ResourceDescription>>> = Default::default();
         let awaiting_connection = Default::default();
         let gateway_public_keys = Default::default();
         let resources_gateways = Default::default();
@@ -281,7 +261,14 @@ where
         registry = register_default_interceptors(registry, &mut media_engine)?;
         let mut setting_engine = SettingEngine::default();
         setting_engine.detach_data_channels();
-        // TODO: Enable UDPMultiplex (had some problems before)
+        setting_engine.set_ip_filter(Box::new({
+            let resources = Arc::clone(&resources);
+            move |ip| !resources.read().values().any(|res_ip| res_ip.contains(ip))
+        }));
+
+        setting_engine.set_interface_filter(Box::new({
+            |name| !name.contains("utun") && name != "tun-firezone"
+        }));
 
         let webrtc_api = APIBuilder::new()
             .with_media_engine(media_engine)
